requirements.txt
# Core Dependencies
fastapi==0.120.0
uvicorn[standard]==0.38.0

# Web Scraping & Rendering
selenium==4.27.1
beautifulsoup4==4.12.3

# HTTP & Utilities
requests==2.32.5
python-dotenv==1.0.1

# HTML Parser
lxml==5.3.0

# Caching
cachetools==5.5.0

# Testing (optional)
pytest==8.3.3
pytest-cov==6.0.0
httpx==0.27.2
